requirements.txt
PyYAML==6.0
langchain==0.0.225
fastapi==0.99.1
google-search-results==2.4.2
openai==0.27.8
streamlit==1.24.0
tiktoken==0.4.0
uvicorn==0.22.0
chromadb==0.3.26
bs4==0.0.1
pymongo==4.4.0
python-dotenv
pybrake
pusher
retry
faiss-cpu